create adding new food for one day process and make some changes in db and style

+++ database/migrations/2025_12_14_191012_create_food_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('food', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('calories');
+            $table->float('calories');
             $table->boolean('is_active')->default(true);
             $table->timestamps();
         });
